Major update to the hybrid workflow class
Realised mistake in that @task and @flow must be high-level routines and
not part of class (got incorrect info from prefect's documentaiton AI).
Have fixed this to have task and flows passed the
HybridQuantumWorkflowBase instance so the flows and tasks need not know
the details of managing flows but can rely on this class to do some of
the management.

Tasks and Flows are back to high-level instances. This breaks the
examples I have and will need to update the examples to make use of the
new class and new task and flow API.

I will copy the older vqpuworkflow.py as vqpuworkflow_deprecated.py in
the next commit and the examples I have that work will make use of this
deprecated interface. I will also rename those to
`${name}_deprecated.py`