Updates to class to add seralization functions
Now the EventClass and the HybridQuantumWorkflowBase classes have
to_dict and from_dict along with a Custom Serialization class that is
used to run the workflows.

This still doesn't work and the issue has to do with the DaskTaskRunner
class that is part of the HybridQuantumWorkflowBase class. This cannot
be trivially serialized and even though the to_dict method does not
explicitly serialize this structure, the presence of this in the
instance causes issues when used with the dask task runners (rather than
the local runner)

A solution might be to remove this from inside the class and just
call the get_dask_taskrunners when necessary in the flows. The idea
would be for the class to have a wrapper function to
get_dask_taskrunners and pick out the appropriate task_runner. Thus, the
class doesn't have to serialize this structure (pity that Prefect didn't
try to provide a serialization method)

I did try adding this to a SillyTestClass that is used in unit tests (to
be included in next commit) and this will be the test to see if this can
work.