Added the QPUMetaData object to store information
This object adds the ability to store metadata about the (v)QPU such as
the number of qubits, the type of qpu, the connectivity, speed, etc.

This object and the data stored will likely grow but it is a good
start point, particularly for adding qpu requirements for a circuit. I
have added this in as well to check that the (v)qpu can run the circuit.
Key will be to add extra functionality to this and better exception
handling so that one can provide multiple qpus as viable options to run
a circuit. I also need to improve how active qpus are managed.